Split style into multiple files.
